nope

+++ src/audioPlayer.js
@@ -36,10 +36,6 @@ class AudioPlayer {
     if (this.currentStream) {
       throw new Error('Bot is currently playing another file on the server.')
     }
-    if (this.ffmpegProcess) {
-      this.ffmpegProcess.kill()
-      delete this.ffmpegProcess
-    }
     await this.join(voiceChannel)
     // Launch the FFMPEG process
     this.ffmpegProcess = spawn(Core.properties.ffmpegBin || 'ffmpeg',
@@ -120,8 +116,10 @@ class AudioPlayer {
    * @param {boolean} disconnect - Set to true to disconnect
    */
   clean (disconnect) {
-    this.ffmpegProcess.kill()
-    delete this.ffmpegProcess
+    if (this.ffmpegProcess) {
+      this.ffmpegProcess.kill()
+      delete this.ffmpegProcess
+    }
     delete this.currentStream
     if (disconnect) {
       try {
